requirements.txt
Django==3.1.7
django-rq==2.4.1
gunicorn==20.1.0 
psycopg2-binary==2.8.6
python-dotenv==0.17.0
rq-scheduler==0.11.0
requests==2.25.1